Fix an issue re the order of NLOCAL in inp_energy_func.
The order of unit in NLOCAL had to be ordered, e.g. NLOCAL(1/2), and
it did not work propery if flipped like NLOCAL(2/1).
This is not how it should work so fixed. Repoted by Max.

This was because flag_nlocal_unit was not formed symmetric in
inp_energy_func, and the double loop in neighbor_assign.F90 run
through only one diagonal (iunit = 1, nunit_all; junit = iunit,
nunit_all).

By this fix, flag_nlocal_unit is now symmetric, so either side can be
referred.